requirements.txt
streamlit~=1.51.0
seaborn~=0.13.2
plotly~=6.4.0
pandas~=2.3.3
geopandas~=1.1.1
numpy~=2.3.4
python-calamine
requests
PyYAML
